% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs_various.R
\name{combMI}
\alias{combMI}
\title{Pool estimes from multiple imputed datasets}
\usage{
combMI(coefmi, varmi, M)
}
\description{
Combines the point and variance estimates obtained from each imputed dataset into the final multiple imputation estimate using Rubin's rules.
}
\keyword{internal}
