% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs_various.R
\name{cumhAdd}
\alias{cumhAdd}
\title{Cumulative hazard for additive model with Weibull baseline.}
\usage{
cumhAdd(tt, ei, K1, K2)
}
\description{
Internal function used to generate data from additive model.
}
\keyword{internal}
