% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs_various.R
\name{cumhCox}
\alias{cumhCox}
\title{Cumulative hazard for Cox model with Weibull baseline.}
\usage{
cumhCox(tt, u, K1, K2)
}
\description{
Internal function used to generate data from Cox model.
}
\keyword{internal}
