% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs_various.R
\name{evMI}
\alias{evMI}
\title{Derive modified event indicator for imputation}
\usage{
evMI(dataset, visit.time, ename)
}
\description{
Used to derive the modified event indicator on the dataset in long format.
}
\keyword{internal}
