% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs_various.R
\name{locf}
\alias{locf}
\title{Apply last observation carried forward to long-format dataset}
\usage{
locf(dataset, marker)
}
\description{
In the dataset in long format, uses the last observation carried forward (LOCF) approach to fill in missing values.
}
\keyword{internal}
