% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs_2Smice.R
\name{mice.impute.lmm}
\alias{mice.impute.lmm}
\title{Draw imputations from a linear mixed model}
\usage{
mice.impute.lmm(ry, namesX, namesZ, nameG, nam, data)
}
\description{
Implements the imputation method "lmm" to specify in mice2 function to draw imputations from linear mixed imputation models.
}
\keyword{internal}
