% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs_2Smice.R
\name{mice2}
\alias{mice2}
\title{Function for multiple imputation by chained equations allowing for "lmm" imputation method.}
\usage{
mice2(data, m = 5, method = vector("character", length = ncol(data)),
  predictorMatrix = (1 - diag(1, ncol(data))),
  visitSequence = (1:ncol(data))[apply(is.na(data), 2, any)],
  form = vector("character", length = ncol(data)),
  post = vector("character", length = ncol(data)), defaultMethod = c("pmm",
  "logreg", "polyreg", "polr"), maxit = 5, diagnostics = TRUE,
  printFlag = TRUE, seed = NA, imputationMethod = NULL,
  defaultImputationMethod = NULL, data.init = NULL, ...)
}
\description{
Modified version of \code{mice} function from the \code{mice} package to allow for "lmm" imputation method.
}
\keyword{internal}
