% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs_2Smice.R
\name{sampler2}
\alias{sampler2}
\title{Internal function for multiple imputation by chained equations allowing for "lmm" imputation method.}
\usage{
sampler2(p, data, m, imp, r, visitSequence, fromto, printFlag, ...)
}
\description{
Modified version of \code{sampler} function from the \code{mice} package to allow for "lmm" imputation method.
}
\keyword{internal}
