% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs_various.R
\name{summ.aalen}
\alias{summ.aalen}
\title{Summary method for semi-parametric additive model fitted with aalen function}
\usage{
summ.aalen(fit)
}
\description{
Extracts hazard differences, confidence intervals and p-values for model fitted using \code{\link[timereg]{aalen}} function.
}
\keyword{internal}
