% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs_various.R
\name{summ.cox}
\alias{summ.cox}
\title{Summary method for Cox model fitted with coxph function}
\usage{
summ.cox(fit)
}
\description{
Extracts hazard ratios, confidence intervals and p-values for model fitted using \code{\link[survival]{coxph}} function.
}
\keyword{internal}
