% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs_various.R
\name{summ.survtd}
\alias{summ.survtd}
\title{Summarise results from MICEa and MICEb methods}
\usage{
summ.survtd(coefs, vars, model, nam, M)
}
\description{
Extracts final multiple imputation estimates of hazard ratios or differences, according to the model, and confidence intervals and p-values
for models fitted using MICEa or the MICEb approach.
}
\keyword{internal}
